api/requirements.txt
Django==5.2.13
django-unfold==0.90.0
django-ninja==1.6.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1
boto3==1.42.91
resend==2.29.0
django-cors-headers==4.9.0
gunicorn==25.3.0
uvicorn==0.44.0
whitenoise==6.12.0
pypdf>=4.0.0
PyMuPDF>=1.24.0
cryptography>=3.1
PyJWT==2.12.1
openai>=1.0.0
stripe>=7.0.0
shortuuid==1.0.13
sentry-sdk[django]>=2.0.0
django-sql-explorer>=5.0.0
django-ratelimit>=4.1.0
django-auditlog
django-q2>=1.6.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
Markdown>=3.5.0
hubspot-api-client>=8.0.0
